agregadas librerías para mejorar un poco más la experiencia de usuario.

+++ js/carrito.js
@@ -81,15 +81,69 @@ function eliminarDelCarrito(e) {
   cargarProductosCarrito(); 
 
   localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+
+  Toastify({
+    text: "Producto eliminado",
+    offset: {
+      x: "1.5rem", // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+      y: "1.5rem" // vertical axis - can be a number or a string indicating unity. eg: '2em'
+    },
+    duration: 3000,
+    /* destination: "https://github.com/apvarun/toastify-js",
+    newWindow: true, */
+    close: true,
+    gravity: "top", // `top` or `bottom`
+    position: "right", // `left`, `center` or `right`
+    stopOnFocus: true, // Prevents dismissing of toast on hover
+    style: {
+      background: "linear-gradient(to right, #4b33a8, #5f36fd)",
+      borderRadius: ".75rem"
+    },
+    onClick: function(){} // Callback after click
+  }).showToast();
   
 }
 
 botonVaciar.addEventListener("click", vaciarCarrito)
 function vaciarCarrito() {
 
-  productosEnCarrito.length = 0;
-  localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
-  cargarProductosCarrito(); 
+  Swal.fire({
+    title: "¿Estás seguro?",
+    icon: "question",
+    html: `Se van a borrar ${productosEnCarrito.reduce((acc, producto) => acc + producto.cantidad, 0)} productos.`,
+    showCancelButton: true,
+    focusConfirm: false,
+    confirmButtonText: `Sí, borrar.`,
+    cancelButtonText: `Mejor no...`,
+  }).then((result) => {
+    /* Read more about isConfirmed, isDenied below */
+    if (result.isConfirmed) {
+      productosEnCarrito.length = 0;
+      localStorage.setItem("productos-en-carrito", JSON.stringify(productosEnCarrito));
+      cargarProductosCarrito();
+
+      Toastify({
+        text: "El carrito se ha vaciado.",
+        offset: {
+          x: "1.5rem", // horizontal axis - can be a number or a string indicating unity. eg: '2em'
+          y: "1.5rem" // vertical axis - can be a number or a string indicating unity. eg: '2em'
+        },
+        duration: 3000,
+        /* destination: "https://github.com/apvarun/toastify-js",
+        newWindow: true, */
+        close: true,
+        gravity: "top", // `top` or `bottom`
+        position: "right", // `left`, `center` or `right`
+        stopOnFocus: true, // Prevents dismissing of toast on hover
+        style: {
+          background: "linear-gradient(to right, #4b33a8, #5f36fd)",
+          borderRadius: ".75rem"
+        },
+        onClick: function(){} // Callback after click
+      }).showToast();
+    }
+  });
+  
 }
 
 function actualizarTotal() {
